amazon_asin_promo_checker/requirements.txt
playwright>=1.44.0
discord.py>=2.3.2
